==> notebooks/03_forecasting_model.ipynb <==
import pandas as pd 
import numpy as np 

df = pd.read_parquet('processed_data/net_flow_model_table.parquet')
df.head()

df['target'].value_counts()

- Can't use train_test_split from sklearn on this data since it's time series data therefore opted for TimeSeriesSplit from sklearn. 
- Make sure to sort data by TIME. 

#split data into training & testing set
from sklearn.model_selection import TimeSeriesSplit 

df = df.sort_values('time_bucket').reset_index(drop = True) 

X = df.drop(columns = ['station_id', 'time_bucket', 'target'])
y = df['target']

tscv = TimeSeriesSplit(n_splits = 3) #3 splits only because 5 seems to crash computer

from sklearn.linear_model import LogisticRegression 
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix

def build_model():
    return LogisticRegression(
        multi_class = 'multinomial',
        max_iter = 500,
        class_weight = 'balanced', 
        n_jobs = -1,
        solver = 'lbfgs'
    )

reports = []

for fold, (train_idx, test_idx) in enumerate(tscv.split(X), start=1):
    X_train, X_test = X.iloc[train_idx], X.iloc[test_idx]
    y_train, y_test = y.iloc[train_idx], y.iloc[test_idx]

    model = build_model()
    model.fit(X_train, y_train)
    
    y_pred = model.predict(X_test)

    print(f'{fold}')
    print(classification_report(y_test, y_pred, digits = 3))

    reports.append(
        classification_report(
            y_test, y_pred, output_dict = True
        )
    )


#aggregate cross-val performance
cv_summary = (
    pd.DataFrame([
        {
            "fold": i + 1,
            "shortage_recall": r["shortage"]["recall"],
            "surplus_recall":  r["surplus"]["recall"],
            "macro_f1":        r["macro avg"]["f1-score"],
        }
        for i, r in enumerate(reports)
    ])
)

print(cv_summary)
cv_summary.mean()


- model correctly identifies over half of all real shortages and surpluses 30 min in advance, using only real net flow history and time features.
- macro f1 relatively low but that's ok because it averages performance across three classes: shortage (rare), surplus (rare) and balanced (dominant). 96.5% of cases are balanced while minority classes are ~1-2% so macro f1 is going be low. 

from sklearn.ensemble import RandomForestClassifier

rf_reports = []

for fold, (train_idx, test_idx) in enumerate(tscv.split(X), start=1):
    X_train, X_test = X.iloc[train_idx], X.iloc[test_idx]
    y_train, y_test = y.iloc[train_idx], y.iloc[test_idx]

    rf = RandomForestClassifier(
        n_estimators=80,
        max_depth=8,
        min_samples_leaf=100,
        class_weight="balanced",
        n_jobs=2,
        random_state=42
    )

    rf.fit(X_train, y_train)
    y_pred = rf.predict(X_test)

    print(f'{fold}')
    print(classification_report(y_test, y_pred, digits=3))

    rf_reports.append(
        classification_report(y_test, y_pred, output_dict=True)
    )


rf_summary = pd.DataFrame([
    {
        "fold": i + 1,
        "shortage_recall": r["shortage"]["recall"],
        "surplus_recall":  r["surplus"]["recall"],
        "macro_f1":        r["macro avg"]["f1-score"],
    }
    for i, r in enumerate(rf_reports)
])

print(rf_summary)
rf_summary.mean()


Compared to Logisitic Regression, the Random Forest saw:
- 4-5pp improvement in shortage recall 
- Macro F1 up by approx 0.07 (relatively big jump in an imbalanced 3 class problem)
- Shortage recall is most important metric because shortages are more operationally costly so the improved shortage recall (53% to 57%) is meaningful. 

label_map = {
    "shortage": 0,
    "balanced": 1,
    "surplus":  2
}

y_enc = df["target"].map(label_map)
y_enc.value_counts()

import xgboost as xgb 

xgb_reports = []

for fold, (train_idx, test_idx) in enumerate(tscv.split(X), start=1):
    X_train, X_test = X.iloc[train_idx], X.iloc[test_idx]
    y_train, y_test = y_enc.iloc[train_idx], y_enc.iloc[test_idx]

    model = xgb.XGBClassifier(
        objective="multi:softprob",
        num_class=3,
        n_estimators=150,        
        max_depth=5,             # shallow trees
        learning_rate=0.1,
        subsample=0.7,
        colsample_bytree=0.7,
        tree_method="hist",      # memory-efficient
        eval_metric="mlogloss",
        n_jobs=2,                # NOT all cores
        random_state=42
    )
    
    model.fit(X_train, y_train)
    y_pred = model.predict(X_test)

    print(f'{fold}')
    print(
        classification_report(
            y_test,
            y_pred,
            target_names=["shortage", "balanced", "surplus"],
            digits=3
        )
    )

    xgb_reports.append(
        classification_report(
            y_test,
            y_pred,
            output_dict=True
        )
    )



xgb_summary = pd.DataFrame([
    {
        "fold": i + 1,
        "shortage_recall": r["0"]["recall"],   # 0 -> shortage
        "surplus_recall":  r["2"]["recall"],   # 2 -> surplus
        "macro_f1":        r["macro avg"]["f1-score"],
    }
    for i, r in enumerate(xgb_reports)
])

print(xgb_summary)
xgb_summary.mean()


- XGBoost Model failed as it fails to predict shortages or surpluses. Instead it overwhelmingly predicts the balanced class.  
- Reject this model. 

best_model = RandomForestClassifier(
    n_estimators=120, #slightly increase n_estimators (from 80)
    max_depth=8,
    min_samples_leaf=100,
    class_weight="balanced",
    n_jobs=2,
    random_state=42
)

best_model.fit(X, y)

#make predictions
rf_preds = df[["station_id", "time_bucket"]].copy()
rf_preds["predicted_status"] = best_model.predict(X)

rf_preds.head()
rf_preds["predicted_status"].value_counts(normalize=True)


from pathlib import Path 

#save predictions
out_dir = Path('processed_data')
out_dir.mkdir(exist_ok = True)

rf_preds.to_parquet(
    out_dir / 'rf_predictions.parquet',
    index = False
)

test = pd.read_parquet("processed_data/rf_predictions.parquet")
test.shape
test["station_id"].nunique()
test["predicted_status"].value_counts(normalize = True)
